refactor(ff-pipeline): extract EncoderConfig into encoder_config.rs
Moves EncoderConfig and EncoderConfigBuilder out of pipeline.rs into a
dedicated encoder_config.rs module, following the single-responsibility
layout used across the rest of the workspace.

Closes #712
