+++ crates/ff-pipeline/src/lib.rs
@@ -42,8 +42,9 @@
 //! ## Module Structure
 //!
 //! - [`audio_pipeline`] — [`AudioPipeline`]
+//! - [`encoder_config`] — [`EncoderConfig`], [`EncoderConfigBuilder`]
 //! - [`error`] — [`PipelineError`]
-//! - [`pipeline`] — [`Pipeline`], [`PipelineBuilder`], [`EncoderConfig`]
+//! - [`pipeline`] — [`Pipeline`], [`PipelineBuilder`]
 //! - [`progress`] — [`Progress`], [`ProgressCallback`]
 //! - [`thumbnail`] — [`ThumbnailPipeline`]
 //! - [`video_pipeline`] — [`VideoPipeline`]
@@ -51,6 +52,7 @@
 #![warn(missing_docs)]
 
 pub mod audio_pipeline;
+pub mod encoder_config;
 pub mod error;
 pub mod pipeline;
 pub mod progress;
@@ -58,8 +60,9 @@ pub mod thumbnail;
 pub mod video_pipeline;
 
 pub use audio_pipeline::AudioPipeline;
+pub use encoder_config::{EncoderConfig, EncoderConfigBuilder};
 pub use error::PipelineError;
-pub use pipeline::{EncoderConfig, EncoderConfigBuilder, Pipeline, PipelineBuilder};
+pub use pipeline::{Pipeline, PipelineBuilder};
 pub use progress::{Progress, ProgressCallback};
 pub use thumbnail::ThumbnailPipeline;
 pub use video_pipeline::VideoPipeline;
