+++ crates/ff-pipeline/src/pipeline.rs
@@ -2,155 +2,19 @@
 //!
 //! This module provides:
 //!
-//! - [`EncoderConfig`] — codec and quality settings for the output file
 //! - [`PipelineBuilder`] — consuming builder that validates configuration
 //! - [`Pipeline`] — the configured pipeline, executed by calling [`run`](Pipeline::run)
 
 use std::time::Instant;
 
 use ff_decode::{AudioDecoder, ImageDecoder, VideoDecoder};
-use ff_encode::{BitrateMode, HardwareEncoder, VideoEncoder};
+use ff_encode::{HardwareEncoder, VideoEncoder};
 use ff_filter::{FilterGraph, HwAccel};
-use ff_format::{AudioCodec, ChapterInfo, Timestamp, VideoCodec};
+use ff_format::{ChapterInfo, Timestamp};
 
+use crate::encoder_config::EncoderConfig;
 use crate::error::PipelineError;
 use crate::progress::{Progress, ProgressCallback};
-
-/// Codec and quality configuration for the pipeline output.
-///
-/// Passed to [`PipelineBuilder::output`] alongside the output path.
-///
-/// Construct via [`EncoderConfig::builder`].
-#[non_exhaustive]
-pub struct EncoderConfig {
-    /// Video codec to use for the output stream.
-    pub video_codec: VideoCodec,
-
-    /// Audio codec to use for the output stream.
-    pub audio_codec: AudioCodec,
-
-    /// Bitrate control mode (CBR, VBR, or CRF).
-    pub bitrate_mode: BitrateMode,
-
-    /// Output resolution as `(width, height)` in pixels.
-    ///
-    /// Resolution precedence in [`Pipeline::run`]:
-    /// 1. This field when `Some` — explicit value always wins.
-    /// 2. The output dimensions of a `scale` filter, inferred automatically.
-    /// 3. The source video's native resolution.
-    ///
-    /// When a `scale` filter is used via [`PipelineBuilder::filter`] you
-    /// typically do **not** need to set this field; the pipeline infers the
-    /// encoder dimensions from the filter. Set it explicitly only to override
-    /// the filter's output size or to resize without a filter.
-    pub resolution: Option<(u32, u32)>,
-
-    /// Output frame rate in frames per second.
-    ///
-    /// `None` preserves the source frame rate.
-    pub framerate: Option<f64>,
-
-    /// Hardware acceleration device to use during encoding.
-    ///
-    /// `None` uses software (CPU) encoding.
-    pub hardware: Option<HwAccel>,
-}
-
-impl EncoderConfig {
-    /// Returns an [`EncoderConfigBuilder`] with sensible defaults:
-    /// H.264 video, AAC audio, CRF 23, no resolution/framerate override, software encoding.
-    #[must_use]
-    pub fn builder() -> EncoderConfigBuilder {
-        EncoderConfigBuilder::new()
-    }
-}
-
-/// Consuming builder for [`EncoderConfig`].
-///
-/// Obtain via [`EncoderConfig::builder`].
-pub struct EncoderConfigBuilder {
-    video_codec: VideoCodec,
-    audio_codec: AudioCodec,
-    bitrate_mode: BitrateMode,
-    resolution: Option<(u32, u32)>,
-    framerate: Option<f64>,
-    hardware: Option<HwAccel>,
-}
-
-impl EncoderConfigBuilder {
-    fn new() -> Self {
-        Self {
-            video_codec: VideoCodec::H264,
-            audio_codec: AudioCodec::Aac,
-            bitrate_mode: BitrateMode::Crf(23),
-            resolution: None,
-            framerate: None,
-            hardware: None,
-        }
-    }
-
-    /// Sets the video codec.
-    #[must_use]
-    pub fn video_codec(mut self, codec: VideoCodec) -> Self {
-        self.video_codec = codec;
-        self
-    }
-
-    /// Sets the audio codec.
-    #[must_use]
-    pub fn audio_codec(mut self, codec: AudioCodec) -> Self {
-        self.audio_codec = codec;
-        self
-    }
-
-    /// Sets the bitrate control mode.
-    #[must_use]
-    pub fn bitrate_mode(mut self, mode: BitrateMode) -> Self {
-        self.bitrate_mode = mode;
-        self
-    }
-
-    /// Convenience: sets `BitrateMode::Crf(crf)`.
-    #[must_use]
-    pub fn crf(mut self, crf: u32) -> Self {
-        self.bitrate_mode = BitrateMode::Crf(crf);
-        self
-    }
-
-    /// Sets the output resolution in pixels.
-    #[must_use]
-    pub fn resolution(mut self, width: u32, height: u32) -> Self {
-        self.resolution = Some((width, height));
-        self
-    }
-
-    /// Sets the output frame rate in frames per second.
-    #[must_use]
-    pub fn framerate(mut self, fps: f64) -> Self {
-        self.framerate = Some(fps);
-        self
-    }
-
-    /// Sets the hardware acceleration backend.
-    #[must_use]
-    pub fn hardware(mut self, hw: HwAccel) -> Self {
-        self.hardware = Some(hw);
-        self
-    }
-
-    /// Builds the [`EncoderConfig`]. Never fails; returns the config directly.
-    #[must_use]
-    pub fn build(self) -> EncoderConfig {
-        EncoderConfig {
-            video_codec: self.video_codec,
-            audio_codec: self.audio_codec,
-            bitrate_mode: self.bitrate_mode,
-            resolution: self.resolution,
-            framerate: self.framerate,
-            hardware: self.hardware,
-        }
-    }
-}
 
 /// A configured, ready-to-run transcode pipeline.
 ///
@@ -561,8 +425,8 @@ impl PipelineBuilder {
     /// Enable two-pass encoding for more accurate bitrate control at a given file size.
     ///
     /// Two-pass encoding is video-only; any audio stream present in the input is
-    /// silently skipped when this flag is set.  Requires [`BitrateMode::Cbr`] or
-    /// [`BitrateMode::Vbr`] on the [`EncoderConfig`].
+    /// silently skipped when this flag is set.  Requires [`ff_encode::BitrateMode::Cbr`] or
+    /// [`ff_encode::BitrateMode::Vbr`] on the [`EncoderConfig`].
     #[must_use]
     pub fn two_pass(mut self) -> Self {
         self.two_pass = true;
